Moved uuidObject.ts from internal to public folder in preparation for exporting it.
update
