+++ change/@microsoft-teams-js-6f5afb78-da58-46be-a695-164d19695294.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "Moved uuidObject.ts from internal to public folder in preparation for exporting it.",
+  "packageName": "@microsoft/teams-js",
+  "email": "erinha@users.noreply.github.com",
+  "dependentChangeType": "patch"
+}
