+++ packages/teams-js/test/internal/communication.spec.ts
@@ -5,9 +5,9 @@ import { MessageRequest } from '../../src/internal/messageObjects';
 import { NestedAppAuthMessageEventNames, NestedAppAuthRequest } from '../../src/internal/nestedAppAuthUtils';
 import { ResponseHandler } from '../../src/internal/responseHandler';
 import { ApiName, ApiVersionNumber, getApiVersionTag } from '../../src/internal/telemetry';
-import { UUID } from '../../src/internal/uuidObject';
 import { ErrorCode, FrameContexts, SdkError } from '../../src/public';
 import * as app from '../../src/public/app/app';
+import { UUID } from '../../src/public/uuidObject';
 import { Utils } from '../utils';
 
 jest.mock('../../src/internal/handlers', () => ({
